requirements.txt
torch
torchvision
kagglehub
opencv-python
numpy
tqdm
scikit-learn
matplotlib
pillow